Rename task status file
