+++ app/drivers/tools/AbstractTool.py
@@ -11,7 +11,7 @@ from app.core import emitter
 from app.core import utilities
 from app.core import values
 from app.core.task.stats import ToolStats
-from app.core.task.status import TaskStatus
+from app.core.task.TaskStatus import TaskStatus
 from app.core.utilities import error_exit
 from app.core.utilities import execute_command
 from app.drivers.AbstractDriver import AbstractDriver
